Delivery Challan module completed. Create page, list out and overview page functions completed.

+++ Finsys/Finsys_App/models.py
@@ -688,7 +688,7 @@ class Fin_Delivery_Challan(models.Model):
     price_list_applied = models.BooleanField(null=True, default=False)
     price_list = models.ForeignKey(Fin_Price_List, on_delete = models.SET_NULL,null=True)
 
-    subtotal = models.BigIntegerField(default=0, null=True)
+    subtotal = models.FloatField(default=0, null=True)
     igst = models.FloatField(default=0.0, null=True, blank=True)
     cgst = models.FloatField(default=0.0, null=True, blank=True)
     sgst = models.FloatField(default=0.0, null=True, blank=True)
